feat(docker): restructure project directories and integrate Docker with PostgreSQL
- Adjusted directory structure: moved API folder to `server/images/api`.
- Added Docker / `docker-compose.yml` for a seamless setup.
- Integrated PostgreSQL database with Docker Compose.
- Validated setup by running `docker-compose up --build` and confirmed API and DB connection on localhost:8080.
