+++ server/.env
@@ -1,0 +1,6 @@
+POSTGRES_USER=example
+POSTGRES_PASSWORD=example
+POSTGRES_DB=test
+POSTGRES_HOST=localhost
+POSTGRES_PORT=5432
+POSTGRES_HOST_AUTH_METHOD=trust
