+++ server/images/api/Dockerfile
@@ -1,0 +1,13 @@
+FROM node:latest
+
+WORKDIR /usr/app
+
+COPY package*.json ./
+
+RUN npm install --quiet
+
+COPY ./ ./
+
+EXPOSE 3000
+
+CMD ["npm", "start"]
