+++ server/images/api/src/index.js
@@ -1,0 +1,38 @@
+const express = require("express");
+const cors = require("cors");
+const { Pool } = require("pg");
+const app = express();
+const port = 3000;
+
+require("dotenv").config();
+
+const pool = new Pool({
+  user: process.env.POSTGRES_USER,
+  host: process.env.POSTGRES_HOST,
+  database: process.env.POSTGRES_DB,
+  password: process.env.POSTGRES_PASSWORD,
+  port: process.env.POSTGRES_PORT,
+});
+
+app.use(express.urlencoded({ extended: true }));
+app.use(express.json());
+app.use(cors());
+
+app.get("/", async (req, res) => {
+  try {
+    const client = await pool.connect();
+    const result = await client.query("SELECT NOW()");
+    client.release();
+    res.json({
+      message: "hellosSss test",
+      timestamp: result.rows[0].now,
+    });
+  } catch (err) {
+    console.error(err);
+    res.send("Error " + err);
+  }
+});
+
+app.listen(port, () => {
+  console.log(`Server listening on ${port}`);
+});
